+++ test/travis_test.exs
@@ -5,10 +5,10 @@ defmodule ExCoveralls.TravisTest do
 
   @content     "defmodule Test do\n  def test do\n  end\nend\n"
   @counts      [0, 1, nil, nil]
-  @source_info [[name: "test/fixtures/test.ex",
+  @source_info [%{name: "test/fixtures/test.ex",
                  source: @content,
                  coverage: @counts
-               ]]
+               }]
 
   test_with_mock "execute", ExCoveralls.Poster, [execute: fn(_) -> "result" end] do
     assert(Travis.execute(@source_info,[]) == "result")
@@ -21,13 +21,22 @@ defmodule ExCoveralls.TravisTest do
     assert(json =~ ~r/source_files/)
   end
 
+  test "generate from env vars" do
+    System.put_env("TRAVIS_BRANCH", "branch")
+    {:ok, payload} = Jason.decode(Travis.generate_json(@source_info))
+    %{"git" =>
+      %{"branch" => branch}
+    } = payload
+    assert(branch == "branch")
+  end
+
   test "submits as `travis-ci` by default" do
-    parsed = Travis.generate_json(@source_info) |> JSX.decode!
+    parsed = Travis.generate_json(@source_info) |> Jason.decode!
     assert(%{ "service_name" => "travis-ci" } = parsed)
   end
 
   test "can be overriden to submit as `travis-pro`" do
-    parsed = Travis.generate_json(@source_info, %{ pro: true, another_key: 3 }) |> JSX.decode!
+    parsed = Travis.generate_json(@source_info, %{ pro: true, another_key: 3 }) |> Jason.decode!
     assert(%{ "service_name" => "travis-pro" } = parsed)
     assert("repo_token" in Map.keys(parsed))
   end